backend/pyproject.toml
[tool.poetry]
name = "qtrial-backend"
version = "0.1.0"
description = "Clinical trial dataset insight agent (Python backend)"
authors = ["Mohamad M Bailoun <mmb84@mail.aub.edu>"]
readme = "README.md"
packages = [{ include = "qtrial_backend", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

pandas = ">=2.1,<4.0.0"
openpyxl = ">=3.1.5,<4.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
typer = ">=0.24.0,<0.25.0"
rich = ">=14.3.2,<15.0.0"
pydantic = ">=2.12.5,<3.0.0"

openai = ">=2.21.0,<3.0.0"
anthropic = ">=0.81.0,<0.82.0"
google-genai = ">=1.63.0,<2.0.0"
scipy = "^1.17.1"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
python-multipart = ">=0.0.9"
boto3 = "^1.42.73"
lifelines = "^0.30.3"
statsmodels = "^0.14.6"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
hypothesis = {extras = ["pandas"], version = ">=6.100.0"}

[tool.poetry.scripts]
qtrial = "qtrial_backend.main:app"

[build-system]
requires = ["poetry-core>=2.3.1,<3.0.0"]
build-backend = "poetry.core.masonry.api"

